+++ custom_components/hacs/blueprints.py
@@ -3,4 +3,5 @@
 from custom_components.hacs.hacsbase import HacsBase
 from custom_components.hacs.hacsrepositorybase import HacsRepositoryBase
 from custom_components.hacs.hacsrepositoryintegration import HacsRepositoryIntegration
+from custom_components.hacs.hacsrepositorybaseplugin import HacsRepositoryPlugin
 from custom_components.hacs.hacsviewbase import HacsViewBase
