+++ custom_components/hacs/hacsrepositoryintegration.py
@@ -1,18 +1,14 @@
 """Blueprint for HacsRepositoryIntegration."""
 # pylint: disable=too-many-instance-attributes,invalid-name,broad-except
-from asyncio import sleep
 from datetime import datetime
 import logging
 import json
-import os
-import shutil
 
-from homeassistant.helpers.event import async_call_later
 from custom_components.hacs.blueprints import HacsRepositoryBase
-from custom_components.hacs.exceptions import HacsBaseException, HacsBlacklistException, HacsMissingManifest
-from custom_components.hacs.handler.download import async_download_file, async_save_file
+from custom_components.hacs.exceptions import HacsBaseException, HacsMissingManifest
 
 _LOGGER = logging.getLogger('custom_components.hacs.repository')
+
 
 class HacsRepositoryIntegration(HacsRepositoryBase):
     """
@@ -29,153 +25,6 @@ class HacsRepositoryIntegration(HacsRepositoryBase):
         self.repository_type = "integration"
         self.manifest_content = None
 
-    async def check_local_directory(self):
-        """Check the local directory."""
-        try:
-            # Remove if it's allready there.
-            if os.path.exists(self.local_path):
-                await self.remove_local_directory()
-
-            # Create the new directory
-            _LOGGER.debug(f"({self.repository_name}) - Creating {self.local_path}")
-            os.mkdir(self.local_path)
-
-        except Exception as exception:
-            _LOGGER.debug(f"({self.repository_name}) - Creating directory {self.local_path} failed with {exception}")
-            return
-
-    async def remove_local_directory(self):
-        """Check the local directory."""
-        try:
-            if os.path.exists(self.local_path):
-                _LOGGER.debug(f"({self.repository_name}) - Removing {self.local_path}")
-                shutil.rmtree(self.local_path)
-
-                while os.path.exists(self.local_path):
-                    _LOGGER.debug(f"({self.repository_name}) - {self.local_path} still exist, waiting 1s and checking again.")
-                    await sleep(1)
-
-        except Exception as exception:
-            _LOGGER.debug(f"({self.repository_name}) - Removing directory {self.local_path} failed with {exception}")
-            return
-
-    async def setup_repository(self):
-        """
-        Run initialation to setup a repository.
-
-        Return True if everything is validated and ok.
-        """
-        try:
-            # Check the blacklist
-            if self.repository_name in self.blacklist:
-                raise HacsBlacklistException
-
-            # If a previous attempt failed we need to reset the track flag
-            self.track = True
-
-            # Set local path
-            if self.local_path is None:
-                integration_name = self.repository_name.split("/")[-1]
-                self.local_path = f"{self.config_dir}/custom_components/{integration_name}"
-
-            # Validate the repository name
-            self.validate_repository_name()
-            
-
-            # Update repository info
-            updateresult = await self.update(False)
-            if not updateresult:
-                self.track = False
-                self.hide = True
-                if self.repository_name not in self.blacklist:
-                    self.blacklist.append(self.repository_name)
-                _LOGGER.debug(f"({self.repository_name}) - Setup failed")
-                return False
-            
-
-        except HacsBaseException as exception:
-            _LOGGER.debug(f"({self.repository_name}) - {exception}")
-            return False
-
-        else:
-            # If we get there all is good.
-            self.start_task_scheduler()
-            if self.repository_id not in self.repositories:
-                self.repositories[self.repository_id] = self
-            _LOGGER.debug(f"({self.repository_name}) - Setup of complete")
-
-            return True
-
-    async def install(self):
-        """Run install tasks."""
-        start_time = datetime.now()
-        _LOGGER.info(f'({self.repository_name}) - Starting installation')
-        try:
-            # Run update
-            await self.update(False)
-            
-
-            # Check local directory
-            await self.check_local_directory()
-            
-
-            # Download files
-            for remote_file in self.content_objects:
-                if remote_file.type == "dir":
-                    continue
-                _LOGGER.debug(f"({self.repository_name}) - Downloading {remote_file.path}")
-
-                filecontent = await async_download_file(self.hass, remote_file.download_url)
-
-                if filecontent is None:
-                    _LOGGER.debug(f"({self.repository_name}) - There was an error downloading the file {remote_file.path}")
-                    continue
-
-                # Save the content of the file.
-                local_file_path = f"{self.local_path}/{remote_file.name}"
-                await async_save_file(local_file_path, filecontent)
-
-
-        except HacsBaseException as exception:
-            _LOGGER.debug(f"({self.repository_name}) - {exception}")
-            return False
-
-        else:
-            self.version_installed = self.last_release_tag
-            self.installed = True
-            self.pending_restart = True
-            _LOGGER.info(f'({self.repository_name}) - installation completed in {(datetime.now() - start_time).seconds} seconds')
-
-    async def remove(self):
-        """Run remove tasks."""
-        from custom_components.hacs.handler.storage import write_to_data_store
-        _LOGGER.debug(f"({self.repository_name}) - Starting removal")
-
-        await self.remove_local_directory()
-
-        if self.repository_id in self.repositories:
-            if not self.installed:
-                del self.repositories[self.repository_id]
-
-        if self.repository_name in self.data["custom"]["integration"]:
-            self.data["custom"]["integration"].remove(self.repository_name)
-
-        write_to_data_store(self.config_dir, self.data)
-
-
-
-    async def uninstall(self):
-        """Run uninstall tasks."""
-        from custom_components.hacs.handler.storage import write_to_data_store
-        _LOGGER.debug(f"({self.repository_name}) - Starting uninstall")
-        await self.remove_local_directory()
-        self.installed = False
-        self.pending_restart = True
-        self.version_installed = None
-        if self.repository_name not in self.data["custom"]["integration"]:
-            del self.repositories[self.repository_id]
-        write_to_data_store(self.config_dir, self.data)
-
     async def update(self, setup=False):
         """Run update tasks."""
         from custom_components.hacs.handler.storage import write_to_data_store
@@ -184,39 +33,9 @@ class HacsRepositoryIntegration(HacsRepositoryBase):
             _LOGGER.info(f'({self.repository_name}) - Starting update')
 
         try:
-            # Set the Gihub repository object
-            self.set_repository()
-
-            # Update description.
-            self.set_description()
-
-            # Set repository ID
-            self.set_repository_id()
-
-            # Set repository releases
-            self.set_repository_releases()
-
-            # Check if last updated string changed.
-            current = self.last_updated
-            new = self.return_last_update()
-            if current == new and current is not None:
-                return True
-            self.last_updated = new
-
-            # Set the repository ref
-            self.set_ref()
-
-            # Set additional info
-            self.set_additional_info()
-
-            # Set repository content
+            self.common_update()
             self.set_repository_content()
-
-            # Set manifest content
             self.set_manifest_content()
-
-            # Run task later
-            self.start_task_scheduler()
 
         except HacsBaseException as exception:
             raise HacsBaseException(exception)
@@ -227,17 +46,9 @@ class HacsRepositoryIntegration(HacsRepositoryBase):
 
         if not setup:
             self.data[self.repository_id] = self
-            write_to_data_store(self.config_dir, self.data)
+            await write_to_data_store(self.config_dir, self.data)
             _LOGGER.info(f'({self.repository_name}) - update completed in {(datetime.now() - start_time).seconds} seconds')
         return True
-
-    def start_task_scheduler(self):
-        """Start task scheduler."""
-        if not self.installed:
-            return
-
-        # Update installed elements every 30min
-        async_call_later(self.hass, 60*30, self.update)
 
     def set_repository_content(self):
         """Set repository content attributes."""
