+++ custom_components/hacs/self.py
@@ -1,0 +1,5 @@
+"""Operations for HACS itself."""
+
+
+def hacs_check_for_update():
+    """Check for updates to HACS."""
